+++ packages/badai-api-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/badai-api-example
 
+## 0.6.44
+
+### Patch Changes
+
+- feat: enhance execution flow and error handling
+- Updated dependencies
+  - @badaitech/badai-api@0.6.44
+
 ## 0.6.43
 
 ### Patch Changes